refactor: clarify public primitive boundaries
Move the MPSC atomic pointer helper under internal and remove the unused public AtomicBox API. Classify primitives in documentation and separate public API integration tests from white-box coverage.
