+++ asyncband/src/mpsc/bounded.rs
@@ -29,8 +29,8 @@ use std::task::Context;
 use std::task::Poll;
 use std::task::Waker;
 
-use crate::atomicbox::AtomicOptionBox;
 use crate::internal::Acquire;
+use crate::internal::AtomicOptionBox;
 use crate::internal::Semaphore;
 use crate::mpsc::RecvError;
 use crate::mpsc::SendError;
