+++ asyncband/src/once/once_cell/mod.rs
@@ -25,9 +25,6 @@ use std::sync::atomic::Ordering;
 use crate::semaphore::Semaphore;
 use crate::semaphore::SemaphorePermit;
 
-#[cfg(test)]
-mod tests;
-
 /// A thread-safe cell which can nominally be written to only once.
 ///
 /// # Examples
