+++ asyncband/tests/once_test/once.rs
@@ -19,10 +19,12 @@ use std::sync::atomic::AtomicUsize;
 use std::sync::atomic::Ordering;
 use std::time::Duration;
 
+use asyncband::once::Once;
 use tokio_test::assert_ready;
 
-use super::*;
-use crate::test_runtime;
+#[path = "../support/runtime.rs"]
+mod support;
+use support::test_runtime;
 
 #[tokio::test]
 async fn test_call_once_runs_only_once() {
