+++ asyncband/tests/once_test/once_cell.rs
@@ -20,9 +20,8 @@ use std::sync::atomic::AtomicBool;
 use std::sync::atomic::Ordering;
 use std::time::Duration;
 
+use asyncband::once::OnceCell;
 use tokio::sync::Mutex;
-
-use super::*;
 
 struct Foo {
     value: Arc<AtomicBool>,
